requirements.txt
python-dotenv
discord
youtube_dl
youtube_search
requests